# Improve bootstrap performance for the scanner module.