… src/compiler/scanner.ts