(rather big) Unicode identifier arrays were initialized no matter if used or not. This change makes sure they are only initialized when they are actually accessed.

In local tests this showed that compilation time when down by 60ms-90ms. (executing `node built/local/tsc.js test.ts`)

<!--
Thank you for submitting a pull request!

Please verify that:
* [-] There is an associated issue in the `Backlog` milestone (**required**)
* [X] Code is up-to-date with the `main` branch
* [X] You've successfully run `gulp runtests` locally
* [-] There are new or updated unit tests validating the change

Refer to CONTRIBUTING.MD for more details.
  https://github.com/Microsoft/TypeScript/blob/main/CONTRIBUTING.md

** Please don't send typo fixes! **
Please don't send a PR solely for the purpose of fixing a typo, unless that
typo truly hurts understanding of the text. Each PR represents work for the
maintainers, and that work should provide commensurate value.

If you're interested in sending a PR, the issue tracker has many issues marked `help wanted`.
-->