@jakebailey 2022-10-21
@typescript-bot perf test this

Though, I expect this to not change as it doesn't really count things like the startup time of the tsc process.