@jakebailey 2022-10-24
I'm working on adding a benchmark for startup times for our outputs, and from what I'm seeing in my testing, this PR doesn't seem to have any affect:

```
main
┌───────────────┬─────────────────┬───────┬───────┐
│ Project       │         Average │  Best │ Worst │
╞═══════════════╧═════════════════╧═══════╧═══════╡
│ tsc --version - node (v16.18.0, x64)            │
├───────────────┬─────────────────┬───────┬───────┤
│ executionTime │ 112ms (± 0.63%) │ 109ms │ 119ms │
╞═══════════════╧═════════════════╧═══════╧═══════╡
│ tsserverlibrary.js load - node (v16.18.0, x64)  │
├───────────────┬─────────────────┬───────┬───────┤
│ executionTime │ 150ms (± 1.02%) │ 146ms │ 180ms │
╞═══════════════╧═════════════════╧═══════╧═══════╡
│ typescript.js load - node (v16.18.0, x64)       │
├───────────────┬─────────────────┬───────┬───────┤
│ executionTime │ 143ms (± 0.66%) │ 138ms │ 153ms │
└───────────────┴─────────────────┴───────┴───────┘

51250
┌───────────────┬─────────────────┬───────┬───────┐
│ Project       │         Average │  Best │ Worst │
╞═══════════════╧═════════════════╧═══════╧═══════╡
│ tsc --version - node (v16.18.0, x64)            │
├───────────────┬─────────────────┬───────┬───────┤
│ executionTime │ 112ms (± 0.89%) │ 108ms │ 121ms │
╞═══════════════╧═════════════════╧═══════╧═══════╡
│ tsserverlibrary.js load - node (v16.18.0, x64)  │
├───────────────┬─────────────────┬───────┬───────┤
│ executionTime │ 149ms (± 0.57%) │ 146ms │ 158ms │
╞═══════════════╧═════════════════╧═══════╧═══════╡
│ typescript.js load - node (v16.18.0, x64)       │
├───────────────┬─────────────────┬───────┬───────┤
│ executionTime │ 143ms (± 0.65%) │ 138ms │ 151ms │
└───────────────┴─────────────────┴───────┴───────┘
```

I'm not sure precisely what you did to test things, but can you try running `tsc --version` locally, maybe?