@marcj 2022-10-24
I also tried `tsc --version` but it had no effect. Only with the command I showed in the first post (`node built/local/tsc.js test.ts`), I got improvements. I guess this is because only an actual compilation/parsing loads the stuff from scanner.ts/parser.ts, but that's just a guess, I'm not familiar with the way code is loaded in tsc.