@marcj 2022-10-24
@jakebailey I pushed the changed to the new memoize function and changed the previous code to use this memoize function instead. A perf test would be interesting, to see if this had any effect.