@jakebailey 2022-10-24
With the way things are executed, the statements you're modifying get unconditionally as they come before the CLI code. It's definitely a bit weird that you're getting that result, but I'll try to see if I can get it to reproduce locally and check it under `pprof` or something.