@RyanCavanaugh 2022-10-24
> In local tests this showed that compilation time when down by 60ms-90ms.

There's just no way this number is right unless this somehow made *parsing itself* faster (or your machine is very very slow) -- the true cold startup time of tsc is barely half that.

Doing three runs of 200 executions each of a minimal tsc run, I get these numbers:

Main: 142.91, 143.75, 145.41
This PR: 144.25, 142.91, 142.82

It's all easily within the margin of error.

Test script
```js
const child_process = require("child_process");
const cmd = "node D:/github/TypeScript-Main/built/local/tsc.js --noEmit --skipLibCheck --noLib D:/throwaway/empty.ts";

test();

async function test() {
    let avg = 0;
    const count = 200;
    for (let i = 0; i < count; i++) {
        avg += await run();
    }
    console.log(avg / count);
}

function run() {
    return new Promise((done) => {
        const start = Date.now();
        child_process.exec(cmd, () => {
            const end = Date.now();
            done(end - start);
        });
    });
}
```

Without a more reliable demonstration of a gain here I don't think we should memoize code that runs once unconditionally.

For kicks I tried a trick I saw once -- since `JSON.parse` has a smaller grammar than JS, you can in theory speed things up by replacing large JSON-able literals with `JSON.parse("[1, 2, 3... ]"). But that doesn't do anything here either - the run time was 144.8, 145.2, 144.7