@jakebailey 2022-10-25
I tested this again and haven't been able to reproduce the speedup. I tried having `test.ts` with contents `console.log("hello")` and saw the same.

Running using hyperfine for statistical results:

```
$ hyperfine --prepare 'git switch main && gulp tsc' --command-name 'main' 'node ./built/local/tsc.js --version' --prepare 'gh pr checkout --detach 51250 && gulp tsc' --command-name 'PR 51250' 'node ./built/local/tsc.js --version'

Benchmark 1: main
  Time (mean ± σ):     117.9 ms ±   1.4 ms    [User: 103.8 ms, System: 13.5 ms]
  Range (min … max):   116.3 ms … 120.4 ms    10 runs
 
Benchmark 2: PR 51250
  Time (mean ± σ):     119.1 ms ±   2.4 ms    [User: 109.4 ms, System: 8.9 ms]
  Range (min … max):   116.2 ms … 124.9 ms    10 runs
 
Summary
  'main' ran
    1.01 ± 0.02 times faster than 'PR 51250'
```

```
$ hyperfine --prepare 'git switch main && gulp tsc' --command-name 'main' 'node ./built/local/tsc.js test.ts' --prepare 'gh pr checkout --detach 51250 && gulp tsc' --command-name 'PR 51250' 'node ./built/local/tsc.js test.ts'  
Benchmark 1: main
  Time (mean ± σ):      2.571 s ±  0.026 s    [User: 4.524 s, System: 0.131 s]
  Range (min … max):    2.525 s …  2.610 s    10 runs
 
Benchmark 2: PR 51250
  Time (mean ± σ):      2.567 s ±  0.017 s    [User: 4.527 s, System: 0.104 s]
  Range (min … max):    2.536 s …  2.601 s    10 runs
 
Summary
  'PR 51250' ran
    1.00 ± 0.01 times faster than 'main'
```