@marcj 2022-10-25
I had the same bootstrap slowdown issues in my C++ port (this is the reason why it caught my eye) which I [fixed in the very same way](https://github.com/marcj/TypeRunner/commit/7eaad414599e82cc94de768f49aafe8720426a69#diff-705e4af5ac7e6af0e5746b8a4707c64ec68295f9a6afd739143a631e72e7fbffR27) and got massive improvements. After quickly hacking together the same in tsc, I got the indicated user time improvements as well (not wall time), but after testing further locally now with hyperfine there is virtually no difference. Seems these optimisation don't work in JS the same way. Not sure why I initially got that improve user time difference. That being said, I also changed in my port regarding `diagnosticInformationMap.generated.ts` and made it lazy too, maybe this improvement was related to that. Weird, closing.