@jakebailey 2022-10-25
@typescript-bot perf test this